cdk/requirements.txt
aws-cdk.core>=1.80.0
aws-cdk.aws_lambda>=1.80.0
aws-cdk.aws_ecr>=1.80.0
botocore
boto3
